Memory leak fixes and debugging changes.
So, it turns out lazily pulling an element from an Array and putting it
back is bad... I think it keeps the array alive.  So in various places
where I was working with the stack or display registers, I was leaking
memory.  Fortunately, some strategic $!'s here and there fixed those
issues.

Then I updated the TRACE opcode to make the argument optional.  Without
an argument, it'll pop the top value off the stack and check that to see
if tracing should be enabled.

Next, I updated the trace output so it's a lot more rich, providing a
view of the display registers, stack, and control registers.

Finally, I modified the CORE opcode to not stop the program.  It simply
shows the program state and then execution proceeds normally.