Moved to IOArray for the stack.
Performance was dismal and I discovered the GC was spending enormous
amounts of time killing old objects which, it seems, were the copies of
the immutable array I was using to model the stack.

So I switched to an IOArray (thus uglifying my nice pure code, but so be
it) and performance is hundreds of times better.

Of course, now there's a small memory leak I need to find...